# backends/catalyst-interviewai/openapi-spec.yaml
swagger: '2.0'
info:
  title: InterviewAI Catalyst Backend API
  description: API for managing user-provided API keys, proxying AI requests, and RAG content processing for InterviewAI.
  version: '2.0.0'
# The host will be automatically generated when you deploy your API Gateway.
# Example: catalyst-gateway-xxxxxx-uc.a.run.app
# You can update this field later after deployment if needed.
host: your-gateway-id.uc.gateway.dev 
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json

# Disable all security - let Cloud Functions handle authentication
security: []
paths:
...
  /api/agents/interview/start:
    options:
      summary: Handle CORS preflight requests for Start Interview
      operationId: corsStartInterview
      security: []
      x-google-backend:
        address: "%s" # Placeholder for Python Agent Gateway URL (21st)
        disable_auth: true
      responses:
        '200':
          description: CORS preflight response
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
    post:
      summary: Starts a new interview session with Python ADK agents
      description: Initializes the orchestrator and agent ecosystem for an interview
      operationId: startInterview
      security: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - interviewType
              - faangLevel
            properties:
              interviewType:
                type: string
                enum: [behavioral, technical, system-design, leadership]
                description: Type of interview to conduct
              faangLevel:
                type: string
                enum: [L3, L4, L5, L6, L7]
                description: Target level for the interview
              resume:
                type: object
                description: Parsed resume data
              jobDescription:
                type: string
                description: Target job description
              targetCompany:
                type: string
                description: Target company name
      x-google-backend:
        address: "%s" # Placeholder for Python Agent Gateway URL (22nd)
        disable_auth: true
      responses:
        '200':
          description: Interview session started successfully
          schema:
            type: object
            properties:
              sessionId:
                type: string
                description: Unique session identifier
              state:
                type: string
                description: Current interview state
              complexity:
                type: string
                description: Assessed complexity level
              initialQuestion:
                type: string
                description: First interview question
              timestamp:
                type: string
                format: date-time
        '400':
          description: Bad request - invalid parameters
        '401':
          description: Unauthorized
        '500':
          description: Failed to start interview session
...
  /api/agents/interview/{sessionId}/end:
    options:
      summary: Handle CORS preflight requests for End Interview
      operationId: corsEndInterview
      security: []
      parameters:
        - name: sessionId
          in: path
          required: true
          type: string
      x-google-backend:
        address: "%s" # Placeholder for Python Agent Gateway URL (27th)
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
      responses:
        '200':
          description: CORS preflight response
          headers:
            Access-Control-Allow-Origin:
              type: string
            Access-Control-Allow-Methods:
              type: string
            Access-Control-Allow-Headers:
              type: string
            Access-Control-Allow-Credentials:
              type: string
    post:
      summary: Ends an interview session and generates final report
      description: Terminates the session and triggers report generation
      operationId: endInterview
      security: []
      parameters:
        - name: sessionId
          in: path
          required: true
          type: string
          description: Interview session ID
      x-google-backend:
        address: "%s" # Placeholder for Python Agent Gateway URL (28th)
        path_translation: APPEND_PATH_TO_ADDRESS
        disable_auth: true
      responses:
        '200':
          description: Interview ended successfully
          schema:
            type: object
            properties:
              success:
                type: boolean
              sessionSummary:
                type: object
                properties:
                  sessionId:
                    type: string
                  totalDuration:
                    type: number
                  finalScores:
                    type: object
                  interventionsCount:
                    type: integer
                  stateTransitionsCount:
                    type: integer
                  complexity:
                    type: string
                  reasoningStrategy:
                    type: string
              message:
                type: string
        '401':
          description: Unauthorized
        '404':
          description: Session not found
        '500':
          description: Failed to end session
...